# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Autenticação, instituições e usuários com múltiplos perfis.",
        "title": "basis-avalia API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/planos/{id}/itens": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "itens-do-plano"
                ],
                "summary": "Adicionar meta ao plano",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.PlanoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            }
        },
        "/api/v1/planos/{id}/itens/{item_id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "itens-do-plano"
                ],
                "summary": "Alterar a quantidade de um item do plano",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.PlanoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "itens-do-plano"
                ],
                "summary": "Remover meta do plano",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.PlanoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.AprovacaoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "orgao": {
                    "type": "string"
                }
            }
        },
...
        "http.CursoEmbutidoResponse": {
            "type": "object",
            "properties": {
                "codigo_emec": {
                    "type": "string"
                },
                "grau": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "modalidade": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "vago": {
                    "type": "boolean"
                }
            }
        },
...
        "http.IndicadorEmbutidoResponse": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "escopo": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "referencia_instrumento": {
                    "type": "string"
                },
                "situacao": {
                    "type": "string"
                }
            }
        },
...
        "http.ItemPlanoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "indicadores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.IndicadorEmbutidoResponse"
                    }
                },
                "meta_id": {
                    "type": "string"
                },
                "meta_nome": {
                    "type": "string"
                },
                "quantidade": {
                    "type": "integer"
                },
                "tem_entrega": {
                    "type": "boolean"
                },
                "versao": {
                    "type": "integer"
                }
            }
        },
...
        "http.PeriodoEmbutidoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
...
        "http.PlanoResponse": {
            "type": "object",
            "properties": {
                "alinhamento_pdi": {
                    "type": "string"
                },
                "alinhamento_ppc": {
                    "type": "string"
                },
                "aprovacao": {
                    "$ref": "#/definitions/http.AprovacaoResponse"
                },
                "atualizado_em": {
                    "type": "string"
                },
                "coordenador_nome": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "curso": {
                    "$ref": "#/definitions/http.CursoEmbutidoResponse"
                },
                "descricao": {
                    "type": "string"
                },
                "encerramento_motivo": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "itens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ItemPlanoResponse"
                    }
                },
                "metas": {
                    "type": "integer"
                },
                "objetivo_geral": {
                    "type": "string"
                },
                "periodo": {
                    "$ref": "#/definitions/http.PeriodoEmbutidoResponse"
                },
                "resultados_esperados": {
                    "type": "string"
                },
                "sem_aprovacao": {
                    "type": "boolean"
                },
                "situacao": {
                    "type": "string"
                },
                "tem_entrega": {
                    "type": "boolean"
                },
                "total_exigido": {
                    "type": "integer"
                },
                "ultimo_documento": {
                    "$ref": "#/definitions/http.UltimoDocumentoResponse"
                },
                "versao": {
                    "type": "integer"
                }
            }
        },
...
        "http.UltimoDocumentoResponse": {
            "type": "object",
            "properties": {
                "gerado_em": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "http.erroResposta": {
            "type": "object",
            "properties": {
                "campo": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}